Criação de diretório e módulo do exercício 108 e sua resolução.

+++ ex107/test.py
@@ -1,5 +1,5 @@
 from ex107 import moeda
-n = int(input('Digite um preço: R$'))
+n = float(input('Digite um preço: R$'))
 print(f'{n} com aumento de 10% é igual a {moeda.aumentar(n, 10)}')
 print(f'{n} menos 10% é igual a {moeda.diminuir(n, 10)}')
 print(f'O dobro de {n} é igual a {moeda.dobro(n)}')
